Split CumUseGraph: pure chart builders to helpers, tooltip to its own module
getDataset and buildChartJSConfig join buildChartConfig in helpers.ts; the
imperative HTML tooltip (renderTooltip/buildTooltipInnerHtml) moves to
chart-tooltip.ts. CumUseGraph.tsx now holds only the chart.js lifecycle
component and theme wrapper. Unit tests cover the pure builders and the
tooltip HTML.
